+++ spike/processing.py
@@ -167,16 +167,32 @@ def comp_sizes(d0,  zflist=None, szmlist=None, largest = LARGESTDATA, sizemin = 
     if (zflist==None) and (szmlist==None):
         zflist=[0]  # do at least something
     szres = []
+
     if (szmlist!=None):         # generate from parameters
         sm1,sm2 = szmlist
-        while True:
-            (si1,si2) = pred_sizes(d0, (sm1,sm2))
-            if debug > 0: print((sm1, sm2))
-            if debug > 0: print((si1, si2))
-            if si1*si2 <= 1.5*SIZEMIN*SIZEMIN:
+        first_loop = True       # 1st loop is specal as there is no downsampling to do
+        while True:      # reduce progressively the multiplier
+            (psi1, psi2) = pred_sizes(d0, (sm1,sm2))   # initial prediction of sizes, ok for FFT
+            if not first_loop:
+                if si1%psi1 != 0:       # check that we'll be able to do downsample
+                    psi1 = si1          # if not, do not reduce
+                if si2%psi2 != 0:       # check that we'll be able to do downsample
+                    psi2 = si2          # if not, do not reduce
+                if psi1 < SIZEMIN//2:          # if too small, 
+                    psi1 = si1           
+                if psi2 < SIZEMIN//2:          # if too small, 
+                    psi2 = si2
+            if debug > 0: print('sm:',(sm1, sm2))
+            if debug > 0: print('psi:',(psi1, psi2))
+            if not first_loop:
+                if (si1,si2) == (psi1, psi2):   # infinite loop
+                    break
+            si1, si2 = psi1, psi2
+            if si1*si2 <= 2*SIZEMIN*SIZEMIN:
                 break
             szres.append( (si1,si2))
             (sm1, sm2) = (0.25*sm1, 0.25*sm2)   # divide by 4, and insure floats
+            first_loop = False
     if (zflist!=None):
         for zf in zflist:
             szres.append( pred_sizes_zf(d0, zf) )
@@ -195,7 +211,7 @@ def comp_sizes(d0,  zflist=None, szmlist=None, largest = LARGESTDATA, sizemin = 
         while sz2 >= sizemin:
             sz2 //= 2
         if not (sz1,sz2) in sizes:
-            sizes.append( (2*sz1, 2*sz2) )
+            sizes.append( (sz1, sz2) )
     if debug>0: print("sizes to process", sizes)
     return sizes
 
@@ -581,9 +597,12 @@ def load_input(name):
 
 class Proc_Parameters(object):
     """this class is a container for processing parameters"""
-    def __init__(self, configfile = None):
-        "initialisation, see processe.mscf for comments on parameters"
-        # processing
+    def __init__(self, configfile=None, verif=True):
+        """initialisation, see process.mscf for comments on parameters
+           if configfile != None, then it should be a config file - parsed with configparser
+           if verif == True (by default) and configfile ! None, the configuration is check for integrity
+        """
+        # processing param
         self.do_F2 = True
         self.do_F1 = True
         self.do_modulus = True
@@ -603,7 +622,7 @@ class Proc_Parameters(object):
         self.szmlist = None
         self.mp = False
         self.nproc = 4
-        # files
+        # files param
         self.apex = None
         self.format = None
         self.infile = None
@@ -618,7 +637,7 @@ class Proc_Parameters(object):
         self.freq_f1demodu = 0.0
         
         if configfile:
-            self.load(configfile)
+            self.load(configfile, verif=verif)
     def from_json(self, jsontxt):
         "updates attributes from json text input"
         dic = json.loads(jsontxt)
@@ -634,8 +653,11 @@ class Proc_Parameters(object):
                 if not callable(v):
                     out[i] =  v
         return json.dumps(out)
-    def load(self, cp):
-        "load from cp config file - should have been opened with ConfigParser() first"
+    def load(self, cp, verif=True):
+        """
+        load from cp config file - should have been opened with ConfigParser() first
+        if verif == True (by default) the configuration is check for integrity
+        """
         if cp.has_option("processing", "sizemulipliers"):   # that nasty bug was around once.
             raise Exception('Error on the name of sizemultiplier parameter, sizemuliplier instead of sizemultiplier')
         self.apex = cp.get( "import", "apex")                                           # input file
@@ -647,6 +669,8 @@ class Proc_Parameters(object):
         self.compress_level = cp.getfloat( "processing", "compress_level", self.compress_level)
         self.tempdir = cp.get( "processing", "tempdir", ".")                            # dir for temporary file
         self.samplingfile = cp.get( "processing", "samplingfile")
+        if self.samplingfile == "None":
+            self.samplingfile = None
         self.samplingfile_fake = cp.getboolean( "processing", "samplingfile_fake", str(self.samplingfile_fake))
         self.largest = cp.getint( "processing", "largest_file", 8*LARGESTDATA)            # largest allowed file
         self.largest = self.largest//8                                                   # in byte in the configfile, internally in word
@@ -682,7 +706,8 @@ class Proc_Parameters(object):
             if debug>0: print("szmlist:", self.szmlist)
         else:
             self.szmlist = None
-        self.verify()
+        if verif:
+            self.verify()
     def verify(self):
         "performs internal coherence of parameters"
         if not self.do_F1 and not self.do_F2:
@@ -706,9 +731,22 @@ class Proc_Parameters(object):
             raise Exception("Please define only one value : zerofilling or sizes multipliers")
         if self.mp and mpiutil.MPI_size > 1:
             raise Exception("use_multiprocessing is not compatible with MPI")
+        if self.samplingfile is not None:
+            if not os.path.exists(self.samplingfile):
+                raise Exception("Sampling file for N.U.S. is missing:",self.samplingfile)
         
     def report(self):
         "print a formatted report"
+        # verify integrity
+        try:
+            self.verify()
+        except:
+            ok = False
+        else:
+            ok = True
+        if not ok:
+            print("------------------- WARNING -------------------------")
+            print(" THIS CONFIGURATION APPEARS TO CONTAIN INCOMPATIBLE PARAMETERS")
         print("------------ processing parameters ------------------")
         for i in dir(self):
             if not i.startswith('_'):
@@ -729,15 +767,16 @@ class Test(unittest.TestCase):
                         #   [16, 17, 18, 19, 20, 21, 22, 23, 24, 25, 26, 27, 28, 29, 30, 31, 32]
     def test_zf(self):
         "testing zerofilling computation"
+        print(self.test_zf.__doc__)
         d = FTICRData(dim = 2)
         d.axis1.size = 1024
         d.axis2.size = 10*1024-10    # 10240 - 10 = 10230
         sizes = comp_sizes(d, zflist=(1,0,-1))
         if SIZEMIN == 1024:
-            self.assertEqual( sizes, [(2048, 20480), (1024, 10240), (1024, 5120), (1024, 1280)])
+            self.assertEqual( sizes, [(2048, 20480), (1024, 10240), (1024, 5120), (512, 640)])
         sizes = comp_sizes(d, szmlist=(3, 1.5) )
         if SIZEMIN == 1024:
-            self.assertEqual( sizes, [(3072, 15360), (1024, 3840), (1024, 1920)])
+            self.assertEqual( sizes, [(3072, 15360), (1024, 3840), (512, 960)])
     def test_proc(self):
         "apply a complete processing test"
         # test is run from above spike
@@ -836,7 +875,11 @@ def main(argv = None):
     #### get parameters from configuration file - store them in a parameter object
     cp = NPKConfigParser()
     print('address configfile is ', configfile)
-    cp.readfp(open(configfile,'r'))
+    with open(configfile,'r') as CF:
+        try:
+            cp.read_file(CF)
+        except:
+            cp.readfp(CF)
     print("reading config file")
     param = Proc_Parameters(cp) # parameters from config file.. 
     # get optionnal parameters
@@ -944,6 +987,7 @@ def main(argv = None):
             print("######################### Checked ################")
     else:
         d1 = None
+        hfar = None
     logflux.log.flush()     # flush logfile
     ###### Do processing
     print("""
@@ -1002,31 +1046,34 @@ def main(argv = None):
     cleaning and closing
 =============================""")
     # copy files and parameters
-    hfar.store_internal_file(filename=configfile, h5name="config.mscf", where='/attached')  # first mscf
-    try:
-        hfar.store_internal_object( h5name='params', obj=d0.hdf5file.retrieve_object(h5name='params') )
-    except:
-        print("No params copied to Output file") 
-    print("parameters and configuration file copied")
-
-    for h5name in ["apexAcquisition.method", "ExciteSweep"]:    # then parameter files
+    if hfar is not None:
+        hfar.store_internal_file(filename=configfile, h5name="config.mscf", where='/attached')  # first mscf
         try:
-            Finh5 = d0.hdf5file.open_internal_file(h5name)
+            hfar.store_internal_object( h5name='params', obj=d0.hdf5file.retrieve_object(h5name='params') )
         except:
-            print("no %s internal file to copy"%h5name)
-        else:   # performed only if no error
-            Fouth5 = hfar.open_internal_file(h5name, access='w')
-            Fouth5.write(Finh5.read())
-            Finh5.close()
-            Fouth5.close()
-            print("%s internal file copied"%h5name)
-    # then logfile
-    logflux.log.flush()     # flush logfile
-    hfar.store_internal_file(filename=logflux.log_name, h5name="processing.log", where='/attached')
-    print("log file copied")
-    # and close
+            print("No params copied to Output file") 
+        else:
+            print("parameters and configuration file copied")
+
+        for h5name in ["apexAcquisition.method", "ExciteSweep"]:    # then parameter files
+            try:
+                Finh5 = d0.hdf5file.open_internal_file(h5name)
+            except:
+                print("no %s internal file to copy"%h5name)
+            else:   # performed only if no error
+                Fouth5 = hfar.open_internal_file(h5name, access='w')
+                Fouth5.write(Finh5.read())
+                Finh5.close()
+                Fouth5.close()
+                print("%s internal file copied"%h5name)
+        # then logfile
+        logflux.log.flush()     # flush logfile
+        hfar.store_internal_file(filename=logflux.log_name, h5name="processing.log", where='/attached')
+        print("log file copied")
+        # and close
+        hfar.close()
     d0.hdf5file.close()
-    hfar.close()
+
     
     if param.mp:
         Pool.close()    # finally closes multiprocessing slaves